requirements.txt
# PSX Research Analyst Dependencies

# HTTP Requests
requests>=2.28.0

# Web Scraping
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Sentiment Analysis
nltk>=3.8.0

# Scheduling
schedule>=1.1.0

# Environment Variables
python-dotenv>=1.0.0

# Progress Bar (optional but nice)
tqdm>=4.64.0

# PDF Processing
PyPDF2>=3.0.0

# Excel Export
openpyxl>=3.1.0
